move all models to models.py
There remained a number of model objects that were in claim.py. This
patch simply consolidates them into models.py.
