+++ models.py
@@ -155,3 +155,107 @@ class Allocation(object):
             self.release_time,
             self.items,
         )
+
+
+class ResourceConstraint(object):
+    def __init__(self, resource_type, min_amount, max_amount,
+                 capability_constraint=None):
+        self.resource_type = resource_type
+        self.min_amount = min_amount
+        self.max_amount = max_amount
+        self.capability_constraint = capability_constraint
+
+    def __repr__(self):
+        return (
+            "ResourceConstraint(resource_type=%s,min_amount=%d,"
+            "max_amount=%d,capabilities=%s)" % (
+                self.resource_type,
+                self.min_amount,
+                self.max_amount,
+                self.capability_constraint,
+            )
+        )
+
+
+class CapabilityConstraint(object):
+    def __init__(self, require_caps=None, forbid_caps=None, any_caps=None):
+        self.require_caps = require_caps
+        self.forbid_caps = forbid_caps
+        self.any_caps = any_caps
+
+    def __repr__(self):
+        return "CapabilityConstraint(require=%s,forbid=%s,any=%s)" % (
+            self.require_caps, self.forbid_caps, self.any_caps,
+        )
+
+
+class ProviderGroupConstraint(object):
+    def __init__(self, require_groups, forbid_groups, any_groups):
+        self.require_groups = require_groups
+        self.forbid_groups = forbid_groups
+        self.any_groups = any_groups
+
+
+class DistanceConstraint(object):
+    def __init__(self, provider, minimum=None, maximum=None):
+        self.provider = provider
+        self.minimum = minimum
+        self.maximum = maximum
+
+
+class ClaimRequestGroupOptions(object):
+    def __init__(self, single_provider=True, isolate_from=None):
+        self.single_provider = single_provider
+        self.isolate_from = isolate_from
+
+
+class ClaimRequestGroup(object):
+    def __init__(self, options=None, resource_constraints=None,
+            capability_constraints=None, provider_group_constraints=None,
+            distance_constraints=None):
+        self.options = options or ClaimRequestGroupOptions()
+        self.resource_constraints = resource_constraints
+        self.capability_constraints = capability_constraints
+        self.provider_group_constraints = provider_group_constraints
+        self.distance_constraints = distance_constraints
+
+
+class ClaimRequest(object):
+    def __init__(self, consumer, request_groups, acquire_time=None,
+            release_time=None):
+        self.consumer = consumer
+        self.request_groups = request_groups
+        self.acquire_time = acquire_time
+        self.release_time = release_time
+
+
+class Claim(object):
+    def __init__(self, acquire_time, release_time, allocation_items,
+            alloc_item_group_map):
+        self.acquire_time = acquire_time
+        self.release_time = release_time
+        self.allocation_items = allocation_items
+        self.allocation_item_to_request_groups = alloc_item_group_map
+
+    def __repr__(self):
+        return "Claim(allocation_items=%s)" % self.allocation_items
+
+
+class Usage(object):
+    def __init__(self, total, reserved, min_unit, max_unit, step_size,
+        allocation_ratio, total_used):
+        self.total = total
+        self.reserved = reserved
+        self.min_unit = min_unit
+        self.max_unit = max_unit
+        self.step_size = step_size
+        self.allocation_ratio = allocation_ratio
+        self.total_used = total_used
+
+
+class ProviderUsages(object):
+    def __init__(self, provider):
+        self.provider = provider
+        # usages is a dict, keyed by resource type internal ID, of Usage
+        # objects
+        self.usages = {}
